requirements.txt
pygame>=2.0.0
PyYAML>=5.4.0
